AVATAR/requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
deepface==0.0.93
diffusers==0.33.1
dill==0.3.8
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.6.0
fsspec==2025.3.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.72.1
gunicorn==23.0.0
h5py==3.14.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
image-similarity-measures==0.3.6
imageio==2.37.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
multidict==6.4.4
multiprocess==0.70.16
namex==0.1.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
phasepack==1.5
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==20.0.0
Pygments==2.19.1
PyMatting==1.1.14
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.3
retina-face==0.0.17
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-image==0.20.0
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.19.0
tifffile==2025.5.10
tokenizers==0.21.1
torch==2.7.0
torcheval==0.0.7
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.2
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
